requirements.txt
asgiref==3.3.4
Django==3.2.3
python-stdnum==1.16
pytz==2021.1
sqlparse==0.4.1
